requirements.txt
Django==4.1.3
django-cors-headers==3.7.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.0.4
jsonschema==4.17.0
mysqlclient==2.1.1
pytz==2022.6
requests==2.28.1